Refactored login and register

+++ Client/App.js
@@ -2,7 +2,8 @@ import React, { useState, useEffect } from 'react';
 import { NavigationContainer } from '@react-navigation/native';
 import { createStackNavigator } from '@react-navigation/stack';
 
-import ApiService from './services/ApiService';
+// import ApiService from './services/ApiService';
+import { registerUserRequest, loginUserRequest, getUserEntries } from './services/ApiService';
 import Home from './screens/home/Home';
 import LoginC from './screens/login/Login';
 import Form from './screens/form/Form';
@@ -14,37 +15,64 @@ import Overview from './screens/overview/Overview';
 import Tabs from './screens/tabs/Tabs';
 
 import config from './config';
+import { Alert } from 'react-native';
 
 const { delHelper, postHelper, dataParser } = config.helperFunctions;
 const Stack = createStackNavigator();
 
 function App() {
-  const [currentUser, setCurrentUser] = useState([]);
+  const [currentUser, setCurrentUser] = useState({});
   const [userEntries, setUserEntries] = useState([]);
 
-  const getUserData = (end, id) => {
-    useEffect(() => {
-      ApiService.profile(end, id).then((data) => {
+  useEffect(() => {
+    if (currentUser.length > 0) {
+      const id = currentUser[0].id;
+      getUserEntries(id).then((data) => {
+        console.log('data', data);
+        console.log('currentUser', currentUser);
         setUserEntries(data[0].entries);
       });
-    }, []);
-  };
+    }
+  }, [currentUser]);
+
+  // const  = (id) => {
+  //   console.log('I get called, outside of use effect');
+  //   useEffect(() => {
+  //     console.log('I get called, inside of use effect');
+  //     getUserEntries(id).then((data) => {
+  //       console.log('data', data);
+  //       console.log('currentUser', currentUser);
+  //       setUserEntries(data[0].entries);
+  //     });
+  //   }, []);
 
   const postOne = (arr, ext, id) => {
     return postHelper(dataParser, arr, ApiService.postOne, setUserEntries, userEntries, ext, id);
   };
 
-  const createUser = (arr, ext) => {
-    return postHelper(dataParser, arr, ApiService.postOne, setCurrentUser, currentUser, ext);
+  // const createUser = (arr, ext) => {
+  //   return postHelper(dataParser, arr, ApiService.postOne, setCurrentUser, currentUser, ext);
+  // };
+
+  const registerUser = async (user) => {
+    const res = await registerUserRequest(user);
+    if (res) {
+      setCurrentUser(res);
+      Alert.alert('User created succesfully');
+    } else {
+      Alert.alert('Username already taken');
+    }
   };
 
-  const postUser = (user) => {
-    console.log('postuser got called from app', user);
-    ApiService.createUser(user);
-    // ApiService.createUser(user).then((data) => {
-    //   console.log('data in app', data);
-    //   setCurrentUser(data);
-    // });
+  const loginUser = async (user) => {
+    const res = await loginUserRequest(user);
+    console.log('loginUser fires');
+    if (res) {
+      setCurrentUser(res);
+      Alert.alert('User logged in succesfully');
+    } else {
+      Alert.alert('Username or password incorrect');
+    }
   };
 
   const deleteOne = (id) => {
@@ -66,26 +94,12 @@ function App() {
 
           <Stack.Screen name="Login">
             {(props) => (
-              <LoginC
-                {...props}
-                createUser={createUser}
-                postUser={postUser}
-                currentUser={currentUser}
-                getUserData={getUserData}
-              />
+              <LoginC {...props} registerUser={registerUser} loginUser={loginUser} currentUser={currentUser} />
             )}
           </Stack.Screen>
 
           <Stack.Screen name="Ui">
-            {(props) => (
-              <Ui
-                {...props}
-                postUser={postUser}
-                userEntries={userEntries}
-                currentUser={currentUser}
-                getUserData={getUserData}
-              />
-            )}
+            {(props) => <Ui {...props} postUser={postUser} userEntries={userEntries} currentUser={currentUser} />}
           </Stack.Screen>
 
           {/* <Stack.Screen name="Form">
@@ -93,7 +107,7 @@ function App() {
               <Form
                 {...props}
                 postOne={postOne}
-                getUserData={getUserData}
+                ={}
                 userEntries={userEntries}
                 currentUser={currentUser}
               />
@@ -102,13 +116,7 @@ function App() {
 
           <Stack.Screen name="Entries">
             {(props) => (
-              <Entries
-                {...props}
-                deleteOne={deleteOne}
-                currentUser={currentUser}
-                getUserData={getUserData}
-                userEntries={userEntries}
-              />
+              <Entries {...props} deleteOne={deleteOne} currentUser={currentUser} userEntries={userEntries} />
             )}
           </Stack.Screen>
